button [nfc]: Move AnimatedScaleOnTap to new file lib/widgets/button.dart
