+++ lib/widgets/button.dart
@@ -1,0 +1,47 @@
+import 'package:flutter/widgets.dart';
+
+/// Apply [Transform.scale] to the child widget when tapped, and reset its scale
+/// when released, while animating the transitions.
+class AnimatedScaleOnTap extends StatefulWidget {
+  const AnimatedScaleOnTap({
+    super.key,
+    required this.scaleEnd,
+    required this.duration,
+    required this.child,
+  });
+
+  /// The terminal scale to animate to.
+  final double scaleEnd;
+
+  /// The duration over which to animate the scale change.
+  final Duration duration;
+
+  final Widget child;
+
+  @override
+  State<AnimatedScaleOnTap> createState() => _AnimatedScaleOnTapState();
+}
+
+class _AnimatedScaleOnTapState extends State<AnimatedScaleOnTap> {
+  double _scale = 1;
+
+  void _changeScale(double scale) {
+    setState(() {
+      _scale = scale;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      behavior: HitTestBehavior.translucent,
+      onTapDown: (_) =>  _changeScale(widget.scaleEnd),
+      onTapUp: (_) =>    _changeScale(1),
+      onTapCancel: () => _changeScale(1),
+      child: AnimatedScale(
+        scale: _scale,
+        duration: widget.duration,
+        curve: Curves.easeOut,
+        child: widget.child));
+  }
+}
