+++ lib/widgets/home.dart
@@ -8,6 +8,7 @@ import 'about_zulip.dart';
 import 'action_sheet.dart';
 import 'app.dart';
 import 'app_bar.dart';
+import 'button.dart';
 import 'color.dart';
 import 'content.dart';
 import 'icons.dart';
@@ -599,51 +600,5 @@ class _AboutZulipButton extends _MenuButton {
   @override
   void onPressed(BuildContext context) {
     Navigator.of(context).push(AboutZulipPage.buildRoute(context));
-  }
-}
-
-/// Apply [Transform.scale] to the child widget when tapped, and reset its scale
-/// when released, while animating the transitions.
-class AnimatedScaleOnTap extends StatefulWidget {
-  const AnimatedScaleOnTap({
-    super.key,
-    required this.scaleEnd,
-    required this.duration,
-    required this.child,
-  });
-
-  /// The terminal scale to animate to.
-  final double scaleEnd;
-
-  /// The duration over which to animate the scale change.
-  final Duration duration;
-
-  final Widget child;
-
-  @override
-  State<AnimatedScaleOnTap> createState() => _AnimatedScaleOnTapState();
-}
-
-class _AnimatedScaleOnTapState extends State<AnimatedScaleOnTap> {
-  double _scale = 1;
-
-  void _changeScale(double scale) {
-    setState(() {
-      _scale = scale;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      behavior: HitTestBehavior.translucent,
-      onTapDown: (_) =>  _changeScale(widget.scaleEnd),
-      onTapUp: (_) =>    _changeScale(1),
-      onTapCancel: () => _changeScale(1),
-      child: AnimatedScale(
-        scale: _scale,
-        duration: widget.duration,
-        curve: Curves.easeOut,
-        child: widget.child));
   }
 }
